+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# 0.0.2
+# 0.0.2 (2018-03-14)
 
 New options: